Renamed util.py to utils.py
Make it match corehq/apps/export/utils.py
