+++ corehq/apps/export/views/edit.py
@@ -24,7 +24,7 @@ from corehq.toggles import MESSAGE_LOG_METADATA, PAGINATED_EXPORTS
 from corehq.apps.export.export import get_export_download, get_export_size
 from corehq.apps.export.models.new import DatePeriod, DataFile, EmailExportWhenDoneRequest
 from corehq.apps.export.views.new import BaseModifyNewCustomView
-from corehq.apps.export.views.util import DailySavedExportMixin, DailySavedExportMixin
+from corehq.apps.export.views.utils import DailySavedExportMixin, DailySavedExportMixin
 from corehq.apps.hqwebapp.views import HQJSONResponseMixin
 from corehq.apps.hqwebapp.utils import format_angular_error, format_angular_success
 from corehq.apps.locations.models import SQLLocation
